Полирнул MVVM
