+++ homework-5/homework-5/DetailedViewMVP+R/Model/FeedDetailedModel.swift
@@ -1,0 +1,16 @@
+//
+//  FeedDetailedModel.swift
+//  homework-5
+//
+//  Created by Vladislav Mishukov on 23.05.2024.
+//
+
+import Foundation
+import UIKit
+
+struct FeedDetailedModel {
+    let title: String
+    let image: UIImage
+    let description: String
+    let sourceLink: String
+}
